Move allButLastAccumulator0 to Parser, rename obj
Organize a few imports, and rename the private object of parser helpers
